requirements.txt
# requirements.txt
pandas==1.4.4
requests==2.28.1
pymongo==4.6.1
python-dotenv==0.21.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
matplotlib==3.5.2
seaborn==0.11.2